Finish book puzzle, Add local space to rail system
The book puzzle is now done. A book is entered into the current order when
pulled, and will return back to it place on its own. When an order is
entered, it will be checked, and an event called for if its correct or
wrong.

The rail system has an option to use local space instead of world space,
making it much easier to work with.