Finish physics ball puzzle and ball machine, Add non-snap grab interactable
The physics ball puzzle requiers the player to move a ball through a maze
to reach the end, where they can get the ball inside. The ball is then
put into a machine that will give them a puzzle piece.

A custom grab interactable has been added, allowing the player to grab
an object at any point without the object being snapped to their hand,
instead being moved relitive to the position and rotation of the hand
when it was grabbed.