Add scene transitions and Implement intro sequence
The transition manager script allows for smooth transitions between scenes
using a fade to black. In order for it to be used, however, the XR rig
has to be moved to a seperate persistant scene. In order to use it, the
persistant scene needs to be the first scene in the heirachy and the
active scene when starting in the editor.

The intro sequence is played before entering the main level, and is started
from the main menu.